Remove Protect Listener changes from this pull request.
Will be parked in another changelist for future consideration
