--- jpos/src/test/java/org/jpos/util/function/ProtectLogEventTest.java
@@ -1,61 +1,0 @@
-/*
- * jPOS Project [http://jpos.org]
- * Copyright (C) 2000-2020 jPOS Software SRL
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package org.jpos.util.function;
-
-import org.jpos.core.Configuration;
-import org.jpos.core.SimpleConfiguration;
-import org.jpos.core.SubConfiguration;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-public class ProtectLogEventTest {
-
-    @Test
-    public void testConstructor() throws Throwable {
-        ProtectLogEvent protectFunction = new ProtectLogEvent();
-        assertNull(protectFunction.wipeFields, "protectedLogListener.wipeFields");
-        assertNull(protectFunction.protectFields, "protectedLogListener.protectFields");
-        assertNull(protectFunction.cfg, "protectedLogListener.cfg");
-    }
-
-    @Test
-    public void testSetConfiguration() throws Throwable {
-        ProtectLogEvent protectFunction = new ProtectLogEvent();
-        Configuration cfg = new SimpleConfiguration();
-        protectFunction.setConfiguration(cfg);
-        assertEquals(0, protectFunction.protectFields.length, "protectedLogListener.protectFields.length");
-        assertEquals(0, protectFunction.wipeFields.length, "protectedLogListener.wipeFields.length");
-        assertSame(cfg, protectFunction.cfg, "protectedLogListener.cfg");
-    }
-
-    @Test
-    public void testSetConfigurationThrowsNullPointerException() throws Throwable {
-        ProtectLogEvent protectFunction = new ProtectLogEvent();
-        Configuration cfg = new SubConfiguration();
-        try {
-            protectFunction.setConfiguration(cfg);
-            fail("Expected NullPointerException to be thrown");
-        } catch (NullPointerException ex) {
-            assertSame(cfg, protectFunction.cfg, "protectedLogListener.cfg");
-            assertNull(ex.getMessage(), "ex.getMessage()");
-            assertNull(protectFunction.wipeFields, "protectedLogListener.wipeFields");
-            assertNull(protectFunction.protectFields, "protectedLogListener.protectFields");
-        }
-    }
-}
